+++ kohii-core/src/main/java/kohii/v1/utils/Capsule.kt
@@ -53,4 +53,6 @@ open class Capsule<T : Any, in A>(
       }
     }
   }
+
+  fun get(arg: A): T = getInstance(arg)
 }