+++ kohii-exoplayer/src/main/java/kohii/v1/exoplayer/TrackSelectorFactory.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2020 Nam Nguyen, nam@ene.im
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package kohii.v1.exoplayer
+
+import android.content.Context
+import com.google.android.exoplayer2.trackselection.DefaultTrackSelector
+
+/**
+ * An interface that is used by the [DefaultExoPlayerProvider] to create a new
+ * [DefaultTrackSelector] when needed.
+ */
+interface TrackSelectorFactory {
+
+  /**
+   * Creates a new [DefaultTrackSelector] instance, given the [Context] of the Application.
+   */
+  fun createDefaultTrackSelector(context: Context): DefaultTrackSelector
+}