+++ kohii-experiments/src/main/java/kohii/v1/experiments/OfficialYouTubePlayerEngine.kt
@@ -36,15 +36,16 @@ class OfficialYouTubePlayerEngine private constructor(
 
   private constructor(context: Context) : this(Master[context])
 
-  companion object : Capsule<OfficialYouTubePlayerEngine, Context>(
-      ::OfficialYouTubePlayerEngine
-  ) {
+  companion object {
+
+    private val capsule =
+      Capsule<OfficialYouTubePlayerEngine, Context>(::OfficialYouTubePlayerEngine)
 
     @JvmStatic // convenient static call for Java
-    operator fun get(context: Context) = super.getInstance(context)
+    operator fun get(context: Context) = capsule.get(context)
 
     @JvmStatic // convenient static call for Java
-    operator fun get(fragment: Fragment) = get(fragment.requireContext())
+    operator fun get(fragment: Fragment) = capsule.get(fragment.requireContext())
   }
 
   override fun prepare(manager: Manager) {