+++ kohii-sample/src/main/java/kohii/v1/sample/ui/overlay/VideoItemHolder.kt
@@ -91,6 +91,7 @@ internal class VideoItemHolder(
   }
 
   override fun onArtworkHint(
+    playback: Playback,
     shouldShow: Boolean,
     position: Long,
     state: Int